Finish api endpoints

+++ exer7/train-tickets/server/mongodb/actions/deleteTicket.js
@@ -5,6 +5,10 @@ export default async function deleteTicket(data) {
     try {
         await connectDB();
         const { ticketID } = data;
+        const existingTicket = await Ticket.findById(ticketID);
+        if (!existingTicket) {
+            return 'Ticket Not Found';
+        }
         await Ticket.findByIdAndDelete(ticketID);
     } catch (e) {
         console.log(e)
